+++ fractal_server/types/__init__.py
@@ -9,8 +9,9 @@ from pydantic.types import StringConstraints
 from fractal_server.urls import normalize_url
 
 from .validators import val_absolute_path
-from .validators import val_canonical_path
 from .validators import val_http_url
+from .validators import val_no_dotdot_in_path
+from .validators import val_non_absolute_path
 from .validators import val_os_path_normpath
 from .validators import val_unique_list
 from .validators import valdict_keys
@@ -29,7 +30,7 @@ A non-empty string, with no leading/trailing whitespaces.
 AbsolutePathStr = Annotated[
     NonEmptyStr,
     AfterValidator(val_absolute_path),
-    AfterValidator(val_canonical_path),
+    AfterValidator(val_no_dotdot_in_path),
     AfterValidator(val_os_path_normpath),
 ]
 """
@@ -39,6 +40,12 @@ Validation fails if the path is not absolute or if it contains a
 parent-directory reference "/../".
 """
 
+RelativePathStr = Annotated[
+    NonEmptyStr,
+    AfterValidator(val_no_dotdot_in_path),
+    AfterValidator(val_os_path_normpath),
+    AfterValidator(val_non_absolute_path),
+]
 
 HttpUrlStr = Annotated[
     NonEmptyStr,
@@ -51,7 +58,7 @@ String representing an URL.
 
 ZarrUrlStr = Annotated[
     NonEmptyStr,
-    AfterValidator(val_canonical_path),
+    AfterValidator(val_no_dotdot_in_path),
     AfterValidator(normalize_url),
 ]
 """
@@ -64,7 +71,7 @@ parent-directory reference "/../".
 
 ZarrDirStr = Annotated[
     NonEmptyStr,
-    AfterValidator(val_canonical_path),
+    AfterValidator(val_no_dotdot_in_path),
     AfterValidator(normalize_url),
 ]
 """